LayoutTests/css3/zoom-coords.xhtml

<html xmlns="http://www.w3.org/1999/xhtml" xmlns:xlink="http://www.w3.org/1999/xlink">
<head>
<link rel="stylesheet" href="../fast/js/resources/js-test-style.css"/>
<style type="text/css">
    body { margin:0px; border:0px; padding:0px; }
    #div1, #svg1 { zoom:200%; }
    #div2, #svg2 { zoom:50%; }
</style>
<script src="../fast/js/resources/js-test-pre.js"></script>
</head>
<body>

<div id="div1" style="width:100px;height:50px;background-color:blue;"></div>
<div id="div2" style="width:200px;height:100px;background-color:green;"></div>
<svg id="svg1" xmlns="http://www.w3.org/2000/svg"
     width="150px" height="50px" viewBox="0 0 150 100"
     preserveAspectRatio="none">
  <rect id="rect1" x="0" y="0" width="100px" height="100px" fill="yellow" />
  <image id="image1" x="100px" y="0" width="50px" height="50px" xlink:href="../svg/dom/resources/magnify.png" />
  <text id="text1" x="100px" y="75px">Test</text>
</svg><svg id="svg2" xmlns="http://www.w3.org/2000/svg"
     width="300px" height="100px" viewBox="0 0 150 100"
     preserveAspectRatio="none">
  <rect id="rect2" x="0" y="0" width="100px" height="100px" fill="orange" />
  <image id="image2" x="100px" y="0" width="50px" height="50px" xlink:href="../svg/dom/resources/magnify.png" />
  <text id="text2" x="100px" y="75px">Test</text>
</svg>

<div id="description"></div>
<div id="console"></div>

<script>

description("This test checks getBoundingClientRect() on zoomed HTML and SVG elements");

debug("Checking HTML elements:");
debug("");

var div1 = document.getElementById("div1").getBoundingClientRect();
shouldBe('div1.left', '0');
shouldBe('div1.top', '0');
shouldBe('div1.width', '100');
shouldBe('div1.height', '50');
shouldBe('div1.right', '100');
shouldBe('div1.bottom', '50');
debug("");

var div2 = document.getElementById("div2").getBoundingClientRect();
shouldBe('div2.left', '0');
shouldBe('div2.top', '200');
shouldBe('div2.width', '200');
shouldBe('div2.height', '100');
shouldBe('div2.right', '200');
shouldBe('div2.bottom', '300');
debug("");

debug("Checking SVG elements:");
debug("");

var svg1 = document.getElementById("svg1").getBoundingClientRect();
shouldBe('svg1.left', '0');
shouldBe('svg1.top', '75');
shouldBe('svg1.width', '150');
shouldBe('svg1.height', '50');
shouldBe('svg1.right', '150');
shouldBe('svg1.bottom', '125');
var rect1 = document.getElementById("rect1").getBoundingClientRect();
shouldBe('rect1.left', '0');
shouldBe('rect1.top', '75');
shouldBe('rect1.width', '100');
shouldBe('rect1.height', '50');
shouldBe('rect1.right', '100');
shouldBe('rect1.bottom', '125');
var image1 = document.getElementById("image1").getBoundingClientRect();
shouldBe('image1.left', '100');
shouldBe('image1.top', '75');
shouldBe('image1.width', '50');
shouldBe('image1.height', '25');
shouldBe('image1.right', '150');
shouldBe('image1.bottom', '100');
var text1 = document.getElementById("text1").getBoundingClientRect();
shouldBe('text1.left == 100', 'true');
shouldBe('text1.top >= 75', 'true');
shouldBe('text1.width > 0', 'true');
shouldBe('text1.height > 0', 'true');
shouldBe('text1.right > 100', 'true');
shouldBe('text1.bottom > 75', 'true');
debug("");

var svg2 = document.getElementById("svg2").getBoundingClientRect();
shouldBe('svg2.left', '600');
shouldBe('svg2.top', '400');
shouldBe('svg2.width', '300');
shouldBe('svg2.height', '100');
shouldBe('svg2.right', '900');
shouldBe('svg2.bottom', '500');
var rect2 = document.getElementById("rect2").getBoundingClientRect();
shouldBe('rect2.left', '600');
shouldBe('rect2.top', '400');
shouldBe('rect2.width', '200');
shouldBe('rect2.height', '100');
shouldBe('rect2.right', '800');
shouldBe('rect2.bottom', '500');
var image2 = document.getElementById("image2").getBoundingClientRect();
shouldBe('image2.left', '800');
shouldBe('image2.top', '400');
shouldBe('image2.width', '100');
shouldBe('image2.height', '50');
shouldBe('image2.right', '900');
shouldBe('image2.bottom', '450');
var text2 = document.getElementById("text2").getBoundingClientRect();
shouldBe('text2.left == 800', 'true');
shouldBe('text2.top >= 400', 'true');
shouldBe('text2.width > 0', 'true');
shouldBe('text2.height > 0', 'true');
shouldBe('text2.right > 800', 'true');
shouldBe('text2.bottom > 400', 'true');
debug("");

successfullyParsed = true;

</script>
<script src="../fast/js/resources/js-test-post.js"></script>

</body>
</html>
