LayoutTests/css3/zoom-coords-expected.txt

Test
Test
This test checks getBoundingClientRect() on zoomed HTML and SVG elements

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Checking HTML elements:

PASS div1.left is 0
PASS div1.top is 0
PASS div1.width is 100
PASS div1.height is 50
PASS div1.right is 100
PASS div1.bottom is 50

PASS div2.left is 0
PASS div2.top is 200
PASS div2.width is 200
PASS div2.height is 100
PASS div2.right is 200
PASS div2.bottom is 300

Checking SVG elements:

PASS svg1.left is 0
PASS svg1.top is 75
PASS svg1.width is 150
PASS svg1.height is 50
PASS svg1.right is 150
PASS svg1.bottom is 125
PASS rect1.left is 0
PASS rect1.top is 75
PASS rect1.width is 100
PASS rect1.height is 50
PASS rect1.right is 100
PASS rect1.bottom is 125
PASS image1.left is 100
PASS image1.top is 75
PASS image1.width is 50
PASS image1.height is 25
PASS image1.right is 150
PASS image1.bottom is 100
PASS text1.left == 100 is true
PASS text1.top >= 75 is true
PASS text1.width > 0 is true
PASS text1.height > 0 is true
PASS text1.right > 100 is true
PASS text1.bottom > 75 is true

PASS svg2.left is 600
PASS svg2.top is 400
PASS svg2.width is 300
PASS svg2.height is 100
PASS svg2.right is 900
PASS svg2.bottom is 500
PASS rect2.left is 600
PASS rect2.top is 400
PASS rect2.width is 200
PASS rect2.height is 100
PASS rect2.right is 800
PASS rect2.bottom is 500
PASS image2.left is 800
PASS image2.top is 400
PASS image2.width is 100
PASS image2.height is 50
PASS image2.right is 900
PASS image2.bottom is 450
PASS text2.left == 800 is true
PASS text2.top >= 400 is true
PASS text2.width > 0 is true
PASS text2.height > 0 is true
PASS text2.right > 800 is true
PASS text2.bottom > 400 is true

PASS successfullyParsed is true

TEST COMPLETE
